+++ ntex/src/web/types/form.rs
@@ -172,7 +172,7 @@ where
 ///     let app = App::new().service(
 ///         web::resource("/index.html")
 ///             // change `Form` extractor configuration
-///             .app_data(
+///             .app_state(
 ///                 web::types::FormConfig::default().limit(4097)
 ///             )
 ///             .route(web::get().to(index))